+++ packages/grid/src/features/standard-table/components/StandardTableCell.tsx
@@ -181,8 +181,6 @@ export const StandardTableCell = React.memo(function StandardTableCell<TItem>({
         zIndex: currentZIndex as CSSProperties["zIndex"],
         height: "var(--current-row-height)",
         background: background,
-        width: width,
-        minWidth: minWidth ?? width ?? "20px",
       }}
     >
       <StandardTableCellUi