+++ packages/grid/src/features/standard-table/components/StandardTableHeadRow.tsx
@@ -183,6 +183,7 @@ export const StandardTableHeadRow = React.memo(function StandardTableHeadRow<
           <Row indent={rowIndent} />
         </th>
       )}
+      <th />
     </TrWithHoverBackground>
   );
 });